+++ spec/puma/plugin/telemetry/formatters/json_formatter_spec.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+module Puma
+  class Plugin
+    module Telemetry
+      module Formatters
+        RSpec.describe JSONFormatter do
+          subject(:formatter) { described_class }
+
+          it 'formats the telemetry as a JSON string' do
+            string = formatter.call('foo' => 'bar')
+
+            data = ::JSON.parse(string)
+            expect(data.fetch('foo')).to eq('bar')
+          end
+
+          it 'handles symbol keys' do
+            string = formatter.call(foo: 'bar')
+
+            data = ::JSON.parse(string)
+            expect(data.fetch('foo')).to eq('bar')
+          end
+        end
+      end
+    end
+  end
+end